loc service swagger update
