--- go-apps/meep-loc-serv/server/inline_response_200_10.go
@@ -1,14 +1,0 @@
-/*
- * Location API
- *
- * The ETSI MEC ISG MEC012 Location API described using OpenAPI. The API is based on the Open Mobile Alliance's specification RESTful Network API for Zonal Presence
- *
- * API version: 1.1.1
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package server
-
-type InlineResponse20010 struct {
-	NotificationSubscriptionList *InlineResponse2001NotificationSubscriptionList `json:"notificationSubscriptionList,omitempty"`
-}
