+++ go-apps/meep-loc-serv/server/model_access_point_info.go
