+++ go-apps/meep-loc-serv/server/model_inline_response_200_3.go
@@ -9,6 +9,6 @@
 
 package server
 
-type InlineResponse20011 struct {
-	NotificationSubscriptionList *InlineResponse2002NotificationSubscriptionList `json:"notificationSubscriptionList,omitempty"`
+type InlineResponse2003 struct {
+	NotificationSubscriptionList *InlineResponse200NotificationSubscriptionList `json:"notificationSubscriptionList,omitempty"`
 }
