+++ go-apps/meep-loc-serv/server/model_location_info.go
