+++ go-apps/meep-loc-serv/server/model_user_event_type.go
