+++ go-apps/meep-loc-serv/server/model_user_info.go
