+++ go-apps/meep-loc-serv/server/model_zonal_presence_notification.go
