+++ go-apps/meep-loc-serv/server/model_zone_info.go
