+++ go-apps/meep-loc-serv/server/model_zone_list.go
