+++ go-apps/meep-loc-serv/server/model_zone_status_notification.go
