--- go-apps/meep-loc-serv/server/subscriptions_api.go
@@ -1,74 +1,0 @@
-/*
- * Location API
- *
- * The ETSI MEC ISG MEC012 Location API described using OpenAPI. The API is based on the Open Mobile Alliance's specification RESTful Network API for Zonal Presence
- *
- * API version: 1.1.1
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package server
-
-import (
-	"net/http"
-)
-
-func UserTrackingSubDelById(w http.ResponseWriter, r *http.Request) {
-	userTrackingSubDelById(w, r)
-}
-
-func UserTrackingSubGet(w http.ResponseWriter, r *http.Request) {
-	userTrackingSubGet(w, r)
-}
-
-func UserTrackingSubGetById(w http.ResponseWriter, r *http.Request) {
-	userTrackingSubGetById(w, r)
-}
-
-func UserTrackingSubPost(w http.ResponseWriter, r *http.Request) {
-	userTrackingSubPost(w, r)
-}
-
-func UserTrackingSubPutById(w http.ResponseWriter, r *http.Request) {
-	userTrackingSubPutById(w, r)
-}
-
-func ZonalTrafficSubDelById(w http.ResponseWriter, r *http.Request) {
-	zonalTrafficSubDelById(w, r)
-}
-
-func ZonalTrafficSubGet(w http.ResponseWriter, r *http.Request) {
-	zonalTrafficSubGet(w, r)
-}
-
-func ZonalTrafficSubGetById(w http.ResponseWriter, r *http.Request) {
-	zonalTrafficSubGetById(w, r)
-}
-
-func ZonalTrafficSubPost(w http.ResponseWriter, r *http.Request) {
-	zonalTrafficSubPost(w, r)
-}
-
-func ZonalTrafficSubPutById(w http.ResponseWriter, r *http.Request) {
-	zonalTrafficSubPutById(w, r)
-}
-
-func ZoneStatusDelById(w http.ResponseWriter, r *http.Request) {
-	zoneStatusDelById(w, r)
-}
-
-func ZoneStatusGet(w http.ResponseWriter, r *http.Request) {
-	zoneStatusGet(w, r)
-}
-
-func ZoneStatusGetById(w http.ResponseWriter, r *http.Request) {
-	zoneStatusGetById(w, r)
-}
-
-func ZoneStatusPost(w http.ResponseWriter, r *http.Request) {
-	zoneStatusPost(w, r)
-}
-
-func ZoneStatusPutById(w http.ResponseWriter, r *http.Request) {
-	zoneStatusPutById(w, r)
-}
